requirements.txt
beautifulsoup4==4.11.1
Flask==2.3.2
Flask_Admin==1.6.1
Flask_Security_Too==5.2.0
flask_sqlalchemy==3.0.3
marko==1.3.0
SQLAlchemy==2.0.12
WTForms==3.0.1
mysqlclient==2.1.1